# custos/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_julesChu12_fly_custos_internal_application_dto.AdminUserInfo:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      last_login_at:
        type: string
      merged_into_user_id:
        type: integer
      nickname:
        type: string
      role:
        type: string
      status:
        type: string
      tenant_id:
        type: integer
      token_version:
        type: integer
      updated_at:
        type: string
      user_type:
        type: string
      username:
        type: string
    type: object
  github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse:
    properties:
      code:
        type: string
      fields:
        additionalProperties: true
        type: object
      message:
        type: string
    type: object
  github_com_julesChu12_fly_custos_internal_application_dto.ForceLogoutUserRequest:
    properties:
      reason:
        maxLength: 500
        type: string
      session_id:
        description: If empty, logout all sessions
        type: string
    type: object
  github_com_julesChu12_fly_custos_internal_application_dto.ForceLogoutUserResponse:
    properties:
      message:
        type: string
      sessions_revoked:
        type: integer
      token_version_new:
        type: integer
      token_version_old:
        type: integer
      user_id:
        type: integer
    type: object
...
  github_com_julesChu12_fly_custos_internal_application_dto.ListUsersResponse:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
      users:
        items:
          $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.AdminUserInfo'
        type: array
    type: object
...
  github_com_julesChu12_fly_custos_internal_application_dto.SystemStatsResponse:
    properties:
      active_sessions:
        type: integer
      active_users:
        type: integer
      deleted_users:
        type: integer
      frozen_users:
        type: integer
      inactive_users:
        type: integer
      new_users_this_week:
        type: integer
      new_users_today:
        type: integer
      total_sessions:
        type: integer
      total_users:
        type: integer
      users_by_role:
        additionalProperties:
          type: integer
        type: object
      users_by_type:
        additionalProperties:
          type: integer
        type: object
    type: object
...
  github_com_julesChu12_fly_custos_internal_application_dto.UpdateUserStatusRequest:
    properties:
      reason:
        description: Optional reason for audit
        maxLength: 500
        type: string
      status:
        enum:
        - active
        - inactive
        - frozen
        - disabled
        - locked
        - deleted
        type: string
    required:
    - status
    type: object
  github_com_julesChu12_fly_custos_internal_application_dto.UpdateUserStatusResponse:
    properties:
      message:
        type: string
      new_status:
        type: string
      old_status:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@fly.com
    name: API Support
    url: http://www.fly.com/support
  description: 用户认证和授权服务API文档，提供用户注册、登录、OAuth、RBAC等功能
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Custos User & Authentication Service API
  version: "1.0"
paths:
  /admin/policies:
    delete:
      consumes:
      - application/json
      description: 删除指定的RBAC策略规则（需要管理员权限）
      parameters:
      - description: 策略规则
        in: body
        name: request
        required: true
        schema:
          properties:
            action:
              type: string
            object:
              type: string
            subject:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除策略规则
      tags:
      - 管理员
    post:
      consumes:
      - application/json
      description: 添加新的RBAC策略规则（需要管理员权限）
      parameters:
      - description: 策略规则
        in: body
        name: request
        required: true
        schema:
          properties:
            action:
              type: string
            object:
              type: string
            subject:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 添加策略规则
      tags:
      - 管理员
  /admin/stats:
    get:
      description: 获取系统统计信息（用户数、会话数等）（需要管理员权限）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.SystemStatsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取系统统计
      tags:
      - 管理员
  /admin/users:
    get:
      description: 获取用户列表，支持分页和筛选（需要管理员权限）
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 用户状态
        in: query
        name: status
        type: string
      - description: 用户角色
        in: query
        name: role
        type: string
      - description: 用户类型
        in: query
        name: user_type
        type: string
      - description: 租户ID
        in: query
        name: tenant_id
        type: integer
      - description: 关键词搜索（用户名、邮箱）
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ListUsersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 用户列表
      tags:
      - 管理员
  /admin/users/{id}:
    get:
      description: 获取指定用户的详细信息（需要管理员权限）
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              active_sessions:
                type: integer
              roles:
                items:
                  type: string
                type: array
              user:
                $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.AdminUserInfo'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取用户详情
      tags:
      - 管理员
  /admin/users/{id}/force-logout:
    post:
      consumes:
      - application/json
      description: 强制指定用户登出（单个会话或全部会话）（需要管理员权限）
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 会话信息（不传则登出所有会话）
        in: body
        name: request
        schema:
          $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ForceLogoutUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ForceLogoutUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 强制用户登出
      tags:
      - 管理员
  /admin/users/{id}/roles:
    get:
      description: 获取指定用户的所有角色和权限（需要管理员权限）
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              permissions:
                items:
                  type: string
                type: array
              roles:
                items:
                  type: string
                type: array
              user_id:
                type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取用户角色
      tags:
      - 管理员
    post:
      consumes:
      - application/json
      description: 为指定用户分配角色（需要管理员权限）
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 角色信息
        in: body
        name: request
        required: true
        schema:
          properties:
            role:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 分配角色
      tags:
      - 管理员
  /admin/users/{id}/status:
    patch:
      consumes:
      - application/json
      description: 更新用户状态（active/inactive/locked/deleted）（需要管理员权限）
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.UpdateUserStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.UpdateUserStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新用户状态
      tags:
      - 管理员
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...