# custos/internal/application/dto/auth.go
package dto

type RegisterRequest struct {
	Username string `json:"username" binding:"required,min=3,max=50"`
	Email    string `json:"email" binding:"required,email"`
	Password string `json:"password" binding:"required,min=8,max=128"`
}

type LoginRequest struct {
	Username string `json:"username" binding:"required"`
	Password string `json:"password" binding:"required"`
}

type LoginMetadata struct {
	IPAddress string
	UserAgent string
}

type RefreshRequest struct {
	SessionID    string `json:"session_id" binding:"required"`
	RefreshToken string `json:"refresh_token" binding:"required"`
}

type LoginResponse struct {
	AccessToken      string    `json:"access_token"`
	TokenType        string    `json:"token_type"`
	ExpiresIn        int64     `json:"expires_in"`
	RefreshToken     string    `json:"refresh_token"`
	RefreshExpiresIn int64     `json:"refresh_expires_in"`
	SessionID        string    `json:"session_id"`
	User             *UserInfo `json:"user"`
}

type UserInfo struct {
	ID       uint   `json:"id"`
	Username string `json:"username,omitempty"`
	Email    string `json:"email,omitempty"`
	Nickname string `json:"nickname,omitempty"`
	Avatar   string `json:"avatar,omitempty"`
	Role     string `json:"role,omitempty"`
	Status   string `json:"status,omitempty"`
}

type ErrorResponse struct {
	Code    string                 `json:"code"`
	Message string                 `json:"message"`
	Fields  map[string]interface{} `json:"fields,omitempty"`
}

type SuccessResponse struct {
	Data interface{} `json:"data"`
}

// OAuth-related DTOs

type OAuthBindRequest struct {
	Code        string `json:"code" binding:"required"`
	State       string `json:"state" binding:"required"`
	RedirectURL string `json:"redirect_url" binding:"required"`
}

type OAuthBindingInfo struct {
	ID          uint   `json:"id"`
	Provider    string `json:"provider"`
	ProviderUID string `json:"provider_uid"`
	BoundAt     string `json:"bound_at"`
}

type OAuthBindingsResponse struct {
	Bindings []OAuthBindingInfo `json:"bindings"`
}

type OAuthBindResponse struct {
	Success bool   `json:"success"`
	Message string `json:"message"`
}

type OAuthUnbindResponse struct {
	Success bool   `json:"success"`
	Message string `json:"message"`
}